Added cape viewer on the cape editor page.

+++ routes/capes/index.js
@@ -1,5 +1,6 @@
 const a = require('express').Router();
 const fetch = require('node-fetch');
+const { Image, createCanvas } = require('canvas');
 const multer = require('multer');
 const gridFS = require('multer-gridfs-storage');
 const Grid = require('gridfs-stream');
@@ -13,15 +14,19 @@ async function checkfName(req, file) {
         return `${file.originalname.replace('.png', '')}.png`;;
     }
 }
-const storage = new gridFS({ url, options: {useUnifiedTopology: true}, file: (req, file) => {    
-    let name = checkfName(req, file).then((n) => {
-        return {
-            bucketName: 'mcmCapes',
-            filename: `${n}`
-        };
-    });
-    return name;
-}});
+const storage = new gridFS({
+    url,
+    options: { useUnifiedTopology: true },
+    file: (req, file) => {
+        let name = checkfName(req, file).then((n) => {
+            return {
+                bucketName: 'mcmCapes',
+                filename: `${n}`
+            };
+        });
+        return name;
+    }
+});
 async function Filter(req, file, cb) {
     if (!file) return cb(new Error(`Could not find file.`));
     if (file.mimetype == "image/png" || file.mimetype == "image/jpeg") {
@@ -43,44 +48,107 @@ const conn = mongoose.createConnection(url, {
 });
 conn.once('open', async function() {
     gridFSBucket = new mongoose.mongo.GridFSBucket(conn.db, { bucketName: 'mcmCapes' });
-    gfs = Grid(conn.db, mongoose.mongo); 
+    gfs = Grid(conn.db, mongoose.mongo);
     gfs.collection('mcmCapes');
 });
 
-a.get('/edit', async function (req, res) {
-    let {theme, auth, capeId} = req.cookies;
+function canvas(image, x, y, w, h) {
+    let version = (image.height >= 64 ? 1 : 0);
+    x = (typeof x === 'undefined' ? 0 : x);
+    y = (typeof y === 'undefined' ? 0 : y);
+    if (version == 0) {
+        w = (typeof w === 'undefined' ? 288 : w);
+        h = (typeof h === 'undefined' ? 288 : h);
+    } else {
+        w = (typeof w === 'undefined' ? image.width * 4.5 : w);
+        h = (typeof h === 'undefined' ? image.height * 4.5 : h);
+    }
+
+    let canvas = createCanvas(w, h);
+    canvas.width = w;
+    canvas.height = h;
+    return canvas;
+}
+
+function capeScale(height) {
+    if (height % 22 === 0) {
+        return height / 22;
+    } else if (height % 17 === 0) {
+        return height / 17;
+    } else if (height >= 32 && (height & (height - 1)) === 0) { // power of 2
+        return height / 32;
+    } else {
+        return Math.max(1, Math.floor(height / 22));
+    }
+}
+
+a.get('/view/:name', async function (req, res) {
+    let { width, height, n } = req.query;
+    if (!width) width = 40;
+    if (!height) height = 64;
+    let image = new Image();
+    image.onload = function() {
+        let op = canvas(image, 0, 0, 269, 256);
+        let context = op.getContext('2d');
+        context.save();
+
+        context.mozImageSmoothingEnabled = false;
+        context.webkitImageSmoothingEnabled = false;
+        context.msImageSmoothingEnabled = false;
+        context.imageSmoothingEnabled = false;
+
+        let cs = capeScale(image.height);
+
+        context.drawImage(image, cs, cs, 10 * cs, 16 * cs, 0, 0, op.width / 2.3, op.height - parseInt(height));
+        context.drawImage(image, 36 * cs, 2 * cs, 11 * cs, 20 * cs, parseInt(width) * 3, 0, op.width / 2, op.height);
+
+        var base64Data = op.toDataURL().replace("data:image/png;base64,", '');
+        var img = Buffer.from(base64Data, 'base64');
+        res.writeHead(200, {
+            'Content-Type': 'image/png',
+            'Content-Length': img.length
+        });
+        res.end(img);
+    }
+    image.onerror = function() { console.error(`Error Loading Cape Image`); res.sendStatus(404); };
+    image.src = `http://api.kaiurl.xyz/capes/${req.params.name}.png`;
+});
+
+a.get('/edit', async function(req, res) {
+    let { theme, auth, capeId } = req.cookies;
     if (!theme) theme = null;
     if (!auth) auth = ""
-    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async (r) => r.json()).then(async (b) => {
+    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async(r) => r.json()).then(async(b) => {
         let findUser = null;
         if (b.OK == true) findUser = await user.findOne({ _id: b.user._id });
         if (b.OK == false && b.code == 12671) return res.redirect('/logout?q=' + req.originalUrl);
         if (!capeId) return res.status(401).render('./error/index', { theme: theme, u: findUser, errorMessage: `Could not authenticate request` });
-        await capes.findOne({ Id: capeId }, async function (e, r) {
-            let {q} = req.query;
+        await capes.findOne({ Id: capeId }, async function(e, r) {
+            let { q } = req.query;
             if (!r || !q) return res.status(401).render('./error/index', { theme: theme, u: findUser, errorMessage: `Could not authenticate request` });
             let findAuth = r.Id.split('-');
             if (findAuth[0] !== q) return res.status(401).render('./error/index', { theme: theme, u: findUser, errorMessage: `Could not authenticate request` });
             res.render('./capes/index', { theme: theme, u: findUser, cape: r });
         });
-    }).catch(e => { console.log(e); res.send('Error') });
+    }).catch(e => { console.log(e);
+        res.send('Error') });
 });
 
-a.get('/edit/auth', async function (req, res) {
-    let {theme, auth} = req.cookies;
+a.get('/edit/auth', async function(req, res) {
+    let { theme, auth } = req.cookies;
     if (!theme) theme = null;
     if (!auth) auth = ""
-    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async (r) => r.json()).then(async (b) => {
+    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async(r) => r.json()).then(async(b) => {
         let findUser = null;
         if (b.OK == true) findUser = await user.findOne({ _id: b.user._id });
         if (b.OK == false && b.code == 12671) return res.redirect('/logout?q=' + req.originalUrl);
-        let {uuid, id} = req.query;
+        let { uuid, id } = req.query;
         if (!uuid || !id) return res.sendStatus(401);
         fetch('https://sessionserver.mojang.com/session/minecraft/profile/' + uuid, {
             method: 'get'
-        }).then((r)=>r.json()).then(async (b)=>{
+        }).then((r) => r.json()).then(async(b) => {
             if (b.error) return console.log(b);
-            await capes.findOne({ uuid: uuid }, async function (e, r) {
+            await capes.findOne({ uuid: uuid }, async function(e, r) {
                 let capeIdAuth = Math.random().toFixed(20).substring(2);
                 if (!r) {
                     new capes({
@@ -91,62 +159,65 @@ a.get('/edit/auth', async function (req, res) {
                         capeId: Math.random(1).toFixed(22).substring(2),
                         linked: false,
                         Id: `${id}-${capeIdAuth}`
-                    }).save().then(()=>{
+                    }).save().then(() => {
                         res.cookie('capeId', `${id}-${capeIdAuth}`);
                         res.redirect('/capes/edit?q=' + id);
                     });
                 } else {
-                    capes.updateOne({ _id: r._id }, { $set: {Id: `${id}-${capeIdAuth}`,cape:`/capes/${b.name}.png`} }).then(()=>{
+                    capes.updateOne({ _id: r._id }, { $set: { Id: `${id}-${capeIdAuth}`, cape: `/capes/${b.name}.png` } }).then(() => {
                         res.cookie('capeId', `${id}-${capeIdAuth}`);
                         res.redirect('/capes/edit?q=' + id);
                     });
                 }
             });
-        }).catch(e=>{res.sendStatus(401)});
-    }).catch(e => { console.log(e); res.send('Error') });
+        }).catch(e => { res.sendStatus(401) });
+    }).catch(e => { console.log(e);
+        res.send('Error') });
 });
 
-a.post('/create', async function (req, res) {
-    let {theme, auth, capeId} = req.cookies;
+a.post('/create', async function(req, res) {
+    let { theme, auth, capeId } = req.cookies;
     if (!theme) theme = null;
     if (!auth) auth = ""
-    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async (r) => r.json()).then(async (b) => {
+    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async(r) => r.json()).then(async(b) => {
         let findUser = null;
         if (b.OK == true) findUser = await user.findOne({ _id: b.user._id });
         if (b.OK == false && b.code == 12671) return res.redirect('/logout?q=' + req.originalUrl);
         if (!capeId) return res.status(401).render('./error/index', { theme: theme, u: findUser, errorMessage: `Could not authenticate request` });
-        await capes.findOne({ Id: capeId }, async function (e, r) {
-            let {q, id} = req.query;
+        await capes.findOne({ Id: capeId }, async function(e, r) {
+            let { q, id } = req.query;
             if (!r || !q || !id) return res.status(401).render('./error/index', { theme: theme, u: findUser, errorMessage: `Could not authenticate request` });
             let findAuth = r.Id.split('-');
             if (findAuth[0] !== q) return res.status(401).render('./error/index', { theme: theme, u: findUser, errorMessage: `Could not authenticate request` });
             if (r.capeId !== id) return res.status(401).render('./error/index', { theme: theme, u: findUser, errorMessage: `Could not authenticate request` });
-            capes.updateOne({ _id: r._id }, { $set: { linked: true, Id: null } }).then(()=> {
+            capes.updateOne({ _id: r._id }, { $set: { linked: true, Id: null } }).then(() => {
                 res.render('./error/index', { theme: theme, u: findUser, errorMessage: `Account is linked. Now press the edit cape button ingame again.` });
             });
         });
-    }).catch(e => { console.log(e); res.send('Error') });
+    }).catch(e => { console.log(e);
+        res.send('Error') });
 });
 
-a.post('/upload/:user/:Id', upload.single('File'), async function (req, res) {
+a.post('/upload/:user/:Id', upload.single('File'), async function(req, res) {
     if (!req.file) return res.status(500).json({ OK: false, error: { message: 'Could not find file' } });
-    let {theme, auth, capeId} = req.cookies;
+    let { theme, auth, capeId } = req.cookies;
     if (!theme) theme = null;
     if (!auth) auth = ""
     if (!capeId) return res.status(401).json({ OK: false, error: { message: `Could not authenticate request` } });
-    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async (r) => r.json()).then(async (b) => {
+    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async(r) => r.json()).then(async(b) => {
         let findUser = null;
         if (b.OK == true) findUser = await user.findOne({ _id: b.user._id });
         if (b.OK == false && b.code == 12671) return res.status(401).json({ OK: false, error: { message: `Could not authenticate request` } });
-        let {q} = req.query; if (!q) return res.status(401).json({ OK: false, error: { message: `Could not authenticate request` } });
-        await capes.findOne({ user: req.params.user, capeId: req.params.Id, Id: capeId }, async function (e, r) {
+        let { q } = req.query;
+        if (!q) return res.status(401).json({ OK: false, error: { message: `Could not authenticate request` } });
+        await capes.findOne({ user: req.params.user, capeId: req.params.Id, Id: capeId }, async function(e, r) {
             if (!r) return res.status(401).json({ OK: false, error: { message: `Could not find cape profile` } });
             fetch('https://api.mojang.com/users/profiles/minecraft/' + r.user, {
                 method: 'get'
-            }).then((r)=>r.json()).then(async (b)=>{
-                await gfs.files.findOne({ uuid: r.uuid }, async function (e, re) {
+            }).then((r) => r.json()).then(async(b) => {
+                await gfs.files.findOne({ uuid: r.uuid }, async function(e, re) {
                     if (!re) {
-                        await gfs.files.findOne({ _id: req.file.id }, async function (e, r) {
+                        await gfs.files.findOne({ _id: req.file.id }, async function(e, r) {
                             if (e) return res.status(500).json({ OK: false, error: e });
                             if (!r) return res.status(500).json({ OK: false, error: 'Could not find File.' });
                             if (r) {
@@ -156,9 +227,9 @@ a.post('/upload/:user/:Id', upload.single('File'), async function (req, res) {
                             }
                         });
                     } else {
-                        gfs.db.collection('mcmCapes.chunks').deleteMany({ files_id: re._id }).then(async () => {
-                            gfs.files.deleteOne({ _id: re._id }).then(async () => {
-                                await gfs.files.findOne({ _id: req.file.id }, async function (e, r) {
+                        gfs.db.collection('mcmCapes.chunks').deleteMany({ files_id: re._id }).then(async() => {
+                            gfs.files.deleteOne({ _id: re._id }).then(async() => {
+                                await gfs.files.findOne({ _id: req.file.id }, async function(e, r) {
                                     if (e) return res.status(500).json({ OK: false, error: e });
                                     if (!r) return res.status(500).json({ OK: false, error: 'Could not find File.' });
                                     if (r) {
@@ -173,30 +244,32 @@ a.post('/upload/:user/:Id', upload.single('File'), async function (req, res) {
                 });
             });
         });
-    }).catch(e => { console.log(e); res.send('Error') });
+    }).catch(e => { console.log(e);
+        res.send('Error') });
 });
 
-a.post('/delete/:user/:Id', async function (req, res) {
-    let {theme, auth, capeId} = req.cookies;
+a.post('/delete/:user/:Id', async function(req, res) {
+    let { theme, auth, capeId } = req.cookies;
     if (!theme) theme = null;
     if (!auth) auth = ""
     if (!capeId) return res.status(401).json({ OK: false, error: { message: `Could not authenticate request` } });
-    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async (r) => r.json()).then(async (b) => {
+    fetch(`http://${req.hostname}/api/auth?q=${auth}`, { method: 'get' }).then(async(r) => r.json()).then(async(b) => {
         let findUser = null;
         if (b.OK == true) findUser = await user.findOne({ _id: b.user._id });
         if (b.OK == false && b.code == 12671) return res.status(401).json({ OK: false, error: { message: `Could not authenticate request` } });
-        let {q} = req.query; if (!q) return res.status(401).json({ OK: false, error: { message: `Could not authenticate request` } });
-        await capes.findOne({ user: req.params.user, capeId: req.params.Id, Id: capeId }, async function (e, r) {
+        let { q } = req.query;
+        if (!q) return res.status(401).json({ OK: false, error: { message: `Could not authenticate request` } });
+        await capes.findOne({ user: req.params.user, capeId: req.params.Id, Id: capeId }, async function(e, r) {
             if (!r) return res.status(401).json({ OK: false, error: { message: `Could not find cape profile` } });
             fetch('https://api.mojang.com/users/profiles/minecraft/' + r.user, {
                 method: 'get'
-            }).then((r)=>r.json()).then(async (b)=>{
-                await gfs.files.findOne({ uuid: r.uuid }, async function (e, re) {
+            }).then((r) => r.json()).then(async(b) => {
+                await gfs.files.findOne({ uuid: r.uuid }, async function(e, re) {
                     if (!re) {
                         res.status(500).json({ OK: false, error: { message: "Could not find cape to remove" } });
                     } else {
-                        gfs.db.collection('mcmCapes.chunks').deleteMany({ files_id: re._id }).then(async () => {
-                            gfs.files.deleteOne({ _id: re._id }).then(async () => {
+                        gfs.db.collection('mcmCapes.chunks').deleteMany({ files_id: re._id }).then(async() => {
+                            gfs.files.deleteOne({ _id: re._id }).then(async() => {
                                 res.redirect(req.headers.referer);
                             }).catch(e => { res.status(500).send(e) });
                         });
@@ -204,10 +277,11 @@ a.post('/delete/:user/:Id', async function (req, res) {
                 });
             });
         });
-    }).catch(e => { console.log(e); res.send('Error') });
+    }).catch(e => { console.log(e);
+        res.send('Error') });
 });
 
-a.get('/:user', async function (req, res) {
+a.get('/:user', async function(req, res) {
     res.redirect('https://api.kaiurl.xyz/capes/' + req.params.user);
 });
